langchain-qdrant-ollama-rag/pyproject.toml
[tool.poetry]
name = "langchain-qdrant-ollama-rag"
version = "0.1.0"
description = "LangChain-Qdrant-Ollama RAG Benchmark with RAGAS Evaluation"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
qdrant-client = "^1.8.0"
ollama = "^0.1.8"
ragas = "^0.1.0"
pydantic = "^2.6.4"
python-dotenv = "^1.0.1"
langchain-community = "^0.0.30"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
